requirements.txt
Django==6.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.14
psycopg2==2.9.11
django-querycount==0.8.3
asgiref==3.11.0
inflection==0.5.1
packaging==26.0
PyJWT==2.11.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
